services/notifier/requirements.txt
python-telegram-bot>=21.0
shared @ file:///app/shared
